multitask fixed

+++ app/gui/web/black_widow/abstract_class/abstract_sniffing_view.py
@@ -28,6 +28,7 @@ import signal
 from app.gui.web.black_widow.abstract_class import AbstractView
 from app.utils.helpers import storage
 from app.utils.helpers import util
+from app.utils.helpers.logger import Log
 
 
 class AbstractSniffingView(AbstractView):
@@ -43,7 +44,10 @@ class AbstractSniffingView(AbstractView):
         """
         pids = storage.read_file(job['pidfile']).split(', ')
         if len(pids) >= 1:
-            return int(pids[0])
+            try:
+                return int(pids[0])
+            except ValueError:
+                pass
         return None
 
     @staticmethod
@@ -52,10 +56,11 @@ class AbstractSniffingView(AbstractView):
         :type job: dict
         """
         pid = AbstractSniffingView._get_job_pid(job)
-        try:
-            os.kill(int(pid), signal.SIGKILL)
-        except ProcessLookupError:
-            pass
+        if pid is not None:
+            try:
+                os.kill(int(pid), signal.SIGKILL)
+            except ProcessLookupError:
+                pass
         storage.delete(job['out_json_file'])
         storage.delete(job['pidfile'])
 
@@ -66,6 +71,8 @@ class AbstractSniffingView(AbstractView):
         """
         session_params = self.session_get(session)
         sniffing_jobs = session_params.get('sniffing_jobs')
+        print('_get_sniffing_jobs:')
+        print(sniffing_jobs)
         update_session = False
         if type(sniffing_jobs) is not dict:
             sniffing_jobs = dict()
@@ -86,6 +93,8 @@ class AbstractSniffingView(AbstractView):
         :type session: django.contrib.sessions.backends.db.SessionStore
         :type sniffing_jobs: dict
         """
+        print('_set_sniffing_jobs:')
+        print(sniffing_jobs)
         session_params = self.session_get(session)
         session_params['sniffing_jobs'] = sniffing_jobs
         self.session_update(session, session_params)
